Moving file locations and fixing module dependencies
