+++ packages/server/src/api/accounts/accounts.module.ts
@@ -18,6 +18,7 @@ import {
 import { Workspaces } from '../workspaces/entities/workspaces.entity';
 import { Organization } from '../organizations/entities/organization.entity';
 import { OrganizationPlan } from '../organizations/entities/organization-plan.entity';
+import { ChannelsModule } from '../channels/channels.module';
 
 @Module({
   imports: [
@@ -35,6 +36,7 @@ import { OrganizationPlan } from '../organizations/entities/organization-plan.en
     forwardRef(() => JourneysModule),
     forwardRef(() => TemplatesModule),
     forwardRef(() => StepsModule),
+    forwardRef(() => ChannelsModule),
     WebhooksModule,
   ],
   controllers: [AccountsController],
