+++ packages/server/src/api/channels/email/interfaces/email.data.ts
