+++ packages/server/src/api/channels/email/providers/resend.provider.ts
@@ -1,6 +1,6 @@
 import { EmailProvider } from '../interfaces/email.provider';
 import { EmailCallbackData, EmailCredentials, EmailProviderData, EmailSendingData, EmailSetupData } from '../interfaces/email.data';
-import { ClickHouseMessage } from '../../../services/clickhouse/interfaces/clickhouse-message';
+import { ClickHouseMessage } from '../../../../common/services/clickhouse/interfaces/clickhouse-message';
 import { ClickHouseEventProvider } from '@/common/services/clickhouse';
 import { ForbiddenException, Injectable } from '@nestjs/common';
 import { BaseLiquidEngineProvider } from '../../interfaces/base.provider';
