+++ packages/server/src/api/channels/email/types/mailgun.data.ts
