+++ packages/server/src/app.module.ts
@@ -63,6 +63,7 @@ import { CacheModule } from '@nestjs/cache-manager';
 import { HealthCheckService } from './app.healthcheck.service';
 import { QueueModule } from '@/common/services/queue/queue.module';
 import { ClickHouseModule } from '@/common/services/clickhouse/clickhouse.module';
+import { ChannelsModule } from './api/channels/channels.module';
 
 const sensitiveKeys = [
   /cookie/i,
@@ -267,6 +268,7 @@ export const formatMongoConnectionString = (mongoConnectionString: string) => {
     RavenModule,
     KafkaModule,
     OrganizationsModule,
+    ChannelsModule,
   ],
   controllers: [AppController],
   providers: getProvidersList(),
